PAHMA-811: Updated queries and cosmetology
Both KIR and OIR can now get and display all three groups of data.
However, the XML in updateKeyInfo has yet to be truly touched, and so
KIR and OIR don't yet actually update any fields except the ones they
already did. (I'm not sure if this was in a previous commit or not, but
they also warn users when they failed to update a field because there
was no match found by autosuggest)